+++ src/main/java/org/toastit_v2/feature/aws/application/service/S3UpLoadServiceImpl.java
@@ -5,6 +5,8 @@ import com.amazonaws.services.s3.model.ObjectMetadata;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+import org.toastit_v2.feature.aws.application.util.FileNameService;
+
 import java.io.IOException;
 import java.util.UUID;
